refactor: update unit tests in HTTP headers factory
