+++ synapse-client-test/src/main/java/io/americanexpress/synapse/client/test/client/BaseClientTest.java
@@ -22,7 +22,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 
 @ExtendWith(SpringExtension.class)
-abstract class BaseClientTest {
+public abstract class BaseClientTest {
 
     protected final XLogger logger = XLoggerFactory.getXLogger(getClass());
