+++ synapse-client-test/src/main/java/io/americanexpress/synapse/client/test/factory/BaseClientHttpHeadersFactoryTest.java
@@ -13,80 +13,79 @@
  */
 package io.americanexpress.synapse.client.test.factory;
 
-import io.americanexpress.synapse.client.client.test.BaseClientTest;
 import io.americanexpress.synapse.client.rest.factory.BaseClientHttpHeadersFactory;
 import io.americanexpress.synapse.client.rest.model.BaseClientRequest;
 import io.americanexpress.synapse.client.rest.model.ClientHeaders;
 import io.americanexpress.synapse.client.rest.model.ClientRouting;
 import io.americanexpress.synapse.client.rest.model.ClientTrace;
+import io.americanexpress.synapse.client.test.client.BaseClientTest;
+import io.americanexpress.synapse.framework.test.CommonAssertionMessages;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpHeaders;
 
 public abstract class BaseClientHttpHeadersFactoryTest<I extends BaseClientRequest,
         H extends BaseClientHttpHeadersFactory<I>> extends BaseClientTest {
 
     @Autowired
-    protected H httpHeadersFactory;
-
-    protected abstract ClientHeaders getDefaultClientHeaders() throws Exception;
-
-    protected abstract I mockDefaultClientRequest();
-
-    public abstract void httpHeadersFactory_constructor();
-
-    public abstract void create_noDependenciesInjected() throws Exception;
-
-    protected abstract void initialize(String url);
+    private H httpHeadersFactory;
 
     protected void setUrl(String url) {
         this.url = url;
     }
-
+    
     @Test
-    public void create_nullArguments() {
-        Assertions.assertThrows(NullPointerException.class, () -> httpHeadersFactory.create(null, null, null));
+    void create_givenNullArguments_expectedNullPointerException() {
+        Assertions.assertThrows(NullPointerException.class, () -> httpHeadersFactory.create(null, null, null), CommonAssertionMessages.EXCEPTION_NOT_THROWN);
     }
 
     @Test
-    public void create_nullClientHeaders() {
-        I request = mockDefaultClientRequest();
-        Assertions.assertThrows(NullPointerException.class, () -> httpHeadersFactory.create(null, request, url));
+    void create_givenNullClientHeaders_expectedNullPointerException() {
+        I clientRequest = mockDefaultClientRequest();
+        Assertions.assertThrows(NullPointerException.class, () -> httpHeadersFactory.create(null, clientRequest, url), CommonAssertionMessages.EXCEPTION_NOT_THROWN);
     }
 
     @Test
-    public void create_nullRequest() {
-        Assertions.assertThrows(NullPointerException.class, () -> httpHeadersFactory.create(new ClientHeaders(), null, url));
+    void create_givenNullClientRequest_expectedNullPointerException() {
+        Assertions.assertThrows(NullPointerException.class, () -> httpHeadersFactory.create(new ClientHeaders(), null, url), CommonAssertionMessages.EXCEPTION_NOT_THROWN);
     }
 
     @Test
-    public void create_emptyClientHeaders() {
-        I request = mockDefaultClientRequest();
-        Assertions.assertThrows(NullPointerException.class, () -> httpHeadersFactory.create(new ClientHeaders(), request, url));
+    void create_givenEmptyClientHeaders_expectedNullPointerException() {
+        I clientRequest = mockDefaultClientRequest();
+        Assertions.assertThrows(NullPointerException.class, () -> httpHeadersFactory.create(new ClientHeaders(), clientRequest, url), CommonAssertionMessages.EXCEPTION_NOT_THROWN);
     }
 
     @Test
-    public void create_emptyClientHeadersRouting() {
+    void create_givenEmptyClientHeadersRouting_expectedNullPointerException() {
         ClientHeaders headers = new ClientHeaders();
         headers.setTrace(new ClientTrace());
-        I request = mockDefaultClientRequest();
-        Assertions.assertThrows(NullPointerException.class, () -> httpHeadersFactory.create(headers, request, url));
+        I clientRequest = mockDefaultClientRequest();
+        Assertions.assertThrows(NullPointerException.class, () -> httpHeadersFactory.create(headers, clientRequest, url), CommonAssertionMessages.EXCEPTION_NOT_THROWN);
     }
 
     @Test
-    public void create_emptyClientHeadersTrace() {
+    void create_givenEmptyClientHeadersTrace_expectedNullPointerException() {
         ClientHeaders headers = new ClientHeaders();
         headers.setRouting(new ClientRouting());
-        I request = mockDefaultClientRequest();
-        Assertions.assertThrows(NullPointerException.class, () -> httpHeadersFactory.create(headers, request, url));
+        I clientRequest = mockDefaultClientRequest();
+        Assertions.assertThrows(NullPointerException.class, () -> httpHeadersFactory.create(headers, clientRequest, url), CommonAssertionMessages.EXCEPTION_NOT_THROWN);
     }
 
     @Test
-    public void create_clean() throws Exception {
+    void create_givenValidClientHeaders_expectedValidHttpHeaders() throws Exception {
         ClientHeaders clientHeaders = getDefaultClientHeaders();
-        I request = mockDefaultClientRequest();
+        I clientRequest = mockDefaultClientRequest();
         String url = "https://example.com";
-        httpHeadersFactory.create(clientHeaders, request, url);
+        HttpHeaders actual = httpHeadersFactory.create(clientHeaders, clientRequest, url);
+        Assertions.assertNotNull(actual, CommonAssertionMessages.VALUE_NOT_NULL);
     }
+    
+    protected abstract void initialize(String url);
+    
+    protected abstract ClientHeaders getDefaultClientHeaders() throws Exception;
+
+    protected abstract I mockDefaultClientRequest();
 }
